+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed scriptable value editor breaking if the value is null
 - Fixed scriptable value editor having the wrong height in newer Unity versions
 - Fixed the package not having an author
+- Fixed allocation when using `foreach` on scriptable lists and dictionaries
 
 ## [1.2.0] - 2023-05-20